trying to get a driver program to work
... to get the functional form of the volume derivative.

I've copied the approach in the examples, but I'm having problems
getting the driver to compile because of the (lack of an) interface to
`forward_model`. In the examples, subroutines are used rather than
functions so that the output type is unnecessary. I feel dodgy about
this, because I don't want to provide volume as a dummy argument to
`forward_model`